Massive cleanup.
