+++ routes/errors.rb
@@ -1,0 +1,14 @@
+require 'sinatra'
+
+module Routing
+    module Errors
+        def self.registered(app)
+              ##
+              # Catch-all 404 error handler.
+              app.not_found do
+                # Display view.
+                slim :not_found
+              end
+        end
+    end
+end
